requirements.txt
Django==1.9.6
djangorestframework==3.3.3
py==1.4.31
pytest==2.9.1
pytest-django==2.9.1
